liquidity selector
